split org module & blueprint into two
